Update id type

+++ src/main/java/assignment2/model/Booking.java
@@ -15,7 +15,7 @@ public class Booking {
 	@Id
 	@Column(name = "id")
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private long id;
+	private Long id;
 
 
 	@CreatedDate
@@ -35,7 +35,7 @@ public class Booking {
 	public Booking() {
 	}
 
-	public Booking(long id, ZonedDateTime dateCreated, String startLocation,
+	public Booking(Long id, ZonedDateTime dateCreated, String startLocation,
 				   String endLocation, ZonedDateTime pickup, ZonedDateTime dropoff,
 				   double distance) {
 		super();
@@ -49,7 +49,7 @@ public class Booking {
 	}
 
 	// Getters
-	public long getId() {
+	public Long getId() {
 		return id;
 	}
 	public ZonedDateTime getDateCreated() {
@@ -72,7 +72,7 @@ public class Booking {
 	}
 
 	// Setters
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public void setDateCreated(ZonedDateTime dateCreated) {
